// src/pages/MakeProposal/types.ts
export interface inputFileTypes {
    inputedLogo: File | null;
    inputedVideo: File | null;
    inputedDocument: File | null;
  }

export interface inputValueTypes {
    name: string;
    description: string;
    email: string;
    github: string;
    social_media: string;
    community: string;
    funding_amount: string;
    equity: string;
    address: string;
}